Create TextureAtlasCreator
Use this script to combine multiple textures and create a texture atlas. 

To Use: 
Execute this script and select existing textured objects within your scene. The UI includes options to name your new object and  new texture, as well as specify the target resolution
and file type of the new texture. You can also specify if you would like to automatically unwrap flat colored objects, and if you would like to optimize UV space in the new texture or 
keep the quilted texture atlas look.

Created textures are automatically applied to the new and old objects with a new material. They are created within the sourceimages folder of your project. 
The original materials and textures are retained. 

Special thanks to George Takahashi. 

Any questions can be directed to spallam42@gmail.com