Turtle Baked Texture Reassigner - Python
This script is written in Python for use with Maya 2016. It was created by Ivan Spalla and published on 1.21.2016. It is free for use under Creative Commons Licensing.

Use this script once you have baked items in a scene using the TURTLE renderer within Maya. Please use the shape flag in the name of your bakes, as this script is heavily dependant on it. Select objects that need to have textures applied, specificy the file folder in which your baked textures can be found, specificy the file prefix, and simply click Assign.

Any questions can be directed to spallam42@gmail.com